made plotting modular
